Add showInLatest attribute
The intention of this feature is to make an item show up in the related
items to the post it refers to, but not show up on the main page. Ie The
main page should have the intial posts, but updates should usually only
show in the related posts to those posts.