Prevent sporadic race condition caused by symlink
Documents are worked on in parallel. Most of the time they are
thread-safe. But when one symlinks to the other, they are not thread
safe because they are writing to the same resources. This can cause
problems ranging from blank/missing content, or double content. Or maybe
something that I haven't thought/seen of yet.

This fix works by explicitly doing symlinks after doing files.
Theoretically, it's possible to re-use the work that's already been
done for the file. However a whole heap of assumptions come with this,
and it's not an expensive operation. So it's better to just let it re-do
it.